requirements.txt

Django==5.2.6
djangorestframework==3.16.1
channels==4.0.0
daphne==4.1.2
psycopg==3.2.10
psycopg-binary==3.2.10
redis==5.0.7
scikit-learn==1.5.1
python-dotenv==1.0.1
pdfminer.six==20231228
markdown-it-py==3.0.0
python-jose[cryptography]==3.3.0
requests==2.32.3
cachetools==5.3.3
tiktoken>=0.7.0
google-generativeai>=0.8.0